python/requirements.txt
streamlit==1.28.0
pandas==2.0.3
numpy==1.24.3
supabase==1.0.3
python-dotenv==1.0.0
plotly==5.17.0
matplotlib==3.7.2
seaborn==0.12.2
scipy==1.10.1
altair==5.0.1
pillow==10.0.0